backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx<0.26,>=0.24
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.4
pytest>=7.4.0
